validando formulário

+++ src/pages/update/Category/index.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
-import { Container, Text } from '../../../components/Loading/style';
 import TemplateBase from '../../../template/TemplateBase';
+import { Container, Text } from '../../../components/Loading/style';
 import Loading from '../../../components/Loading';
 import caterogiesRepository from '../../../repositories/categories';
 import useForm from '../../../hooks/useForm';
@@ -11,6 +11,21 @@ import Button from '../../../components/Button';
 
 const H1 = styled.h1`
 color:var(--grayDark);`;
+
+export const ContainerMiniLoading = styled.div`
+display:flex;
+justify-content:space-between;
+align-items: center;
+width:73%;
+@media(min-width: 768px) {
+    width:25%;
+  }
+`;
+
+export const TextMiniLoading = styled.span`
+display:block;
+color:var(--grayDark);
+`;
 
 const Category = () => {
   const category = {
@@ -43,8 +58,31 @@ const Category = () => {
   }, []);
 
   const deleteCategory = () => {
-    console.log('vou exluir: ', categoryToEdit);
+    if (Object.keys(categoryToEdit).length === 0) {
+      console.log('Selecione uma Categoria para excluir');
+      return;
+    }
     caterogiesRepository.deleteCategory(categoryToEdit).then((updatedCategories) => {
+      setCategories([...categories, updatedCategories]);
+      clearForm();
+      setIsFetching(false);
+    }).catch((err) => setIsFetching(true));
+  };
+
+  const editCategory = (event) => {
+    event.preventDefault();
+    const validCategory = categories.find((categoryToCheck) => categoryToCheck.title === values.title);
+    if (!validCategory || validCategory.length === 0) {
+      console.log('Edite uma categoria válida');
+      return;
+    }
+
+    setIsFetching(true);
+
+    const editedCategory = { title: values.title, description: values.description, color: values.color };
+    const categoryToSend = { ...categoryToEdit, ...editedCategory };
+    console.log(categoryToSend);
+    caterogiesRepository.update(categoryToSend).then((updatedCategories) => {
       setCategories([...categories, updatedCategories]);
       clearForm();
       setIsFetching(false);
@@ -59,7 +97,7 @@ const Category = () => {
       </Container>
       )}
 
-      { categories.length >= 1 && (<Select onChange={getCategory} categories={categories} />) }
+      {categories.length >= 1 && (<Select onChange={getCategory} categories={categories} />)}
 
       <H1>
         Editar Categoria:
@@ -67,42 +105,31 @@ const Category = () => {
         {values.title}
       </H1>
 
-      <form onSubmit={(e) => {
-        e.preventDefault();
-        setIsFetching(true);
-        const editedCategory = { title: values.title, description: values.description, color: values.color };
-        const categoryToSend = { ...categoryToEdit, ...editedCategory };
-        caterogiesRepository.update(categoryToSend).then((updatedCategories) => {
-          setCategories([...categories, updatedCategories]);
-          clearForm();
-          setIsFetching(false);
-        }).catch((err) => setIsFetching(true));
-      }}
-      >
-        <FormField label="Título da Categoria" type="text" name="title" value={values.title} onChange={setValue} />
-        <FormField label="Descrição" type="textarea" name="description" value={values.description} onChange={setValue} />
+      <form onSubmit={(e) => editCategory(e)}>
+        <FormField label="Título da Categoria" type="text" name="title" value={values.title} onChange={setValue} required />
+        <FormField label="Descrição" type="textarea" name="description" value={values.description} onChange={setValue} required />
         <FormField label="Cor" type="color" name="color" value={values.color} onChange={setValue} />
         {
-      isFetching && (
-      <Container>
-        <Loading />
-        <Text>editando categoria...</Text>
-      </Container>
-      )
-     }
+     isFetching && (
+     <ContainerMiniLoading>
+       <Loading />
+       <TextMiniLoading>editando categoria...</TextMiniLoading>
+     </ContainerMiniLoading>
+     )
+    }
         {
-      !isFetching && (
-      <div>
-        <Button type="button" onClick={() => deleteCategory()} background="#F15E5E">
-          Excluir Categoria
-        </Button>
-        <Button>
-          Editar
-        </Button>
-      </div>
+     !isFetching && (
+     <div>
+       <Button type="button" onClick={() => deleteCategory()} background="#F15E5E">
+         Excluir Categoria
+       </Button>
+       <Button>
+         Editar
+       </Button>
+     </div>
 
-      )
-     }
+     )
+    }
       </form>
     </TemplateBase>
   );
